feat(rust): add `reset` command, refactor `node delete`
